Major Design Update
- Created Upload Error Page
- Merged upload and dashboard app
- Excluded tags for working environment
- Tested whole app
