+++ atlasDB/atlas/templates/atlas/layout.html
@@ -12,7 +12,7 @@
 		{% block content %}
 			<ul class="collection">
 				{% for task in all_tasks %}
-					<li class="collection-item"><a href="aufgabe/{{task.id}}">{{ task.titel }} ({{ task.lehrer }})</a> - <a href="{{ task.datensatz }}">Download</a></li>
+					<li class="collection-item"><a href="/aufgabe/{{ task.id }}">{{ task.titel }} ({{ task.lehrer }})</a> - <a href="{{ task.datensatz }}">Download</a></li>
 				{% endfor %}
 			</ul>
 		{% endblock %}
