+++ atlasDB/atlas/templates/atlas/upload/error.html
@@ -1,0 +1,1 @@
+error
