+++ atlasDB/atlas/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path, include
+from . import views
+
+urlpatterns = [
+	path('', views.upload),
+	path('upload/', views.upload, name='upload'),
+	path('upload/success/', views.success, name='success'),
+	path('upload/error/', views.error, name='error'),
+	path('dashboard/', views.dashboard, name='dashboard'),
+	path('aufgabe/<int:id>/', views.aufgabe, name='aufgabe'),
+]
