+++ atlasDB/atlasDB/settings.py
@@ -18,8 +18,7 @@ ALLOWED_HOSTS = []
 
 INSTALLED_APPS = [
     'changelog',
-    'upload',
-    'dashboard',
+    'atlas',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
