--- atlasDB/dashboard/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class DashboardConfig(AppConfig):
-    name = 'dashboard'
