--- atlasDB/dashboard/static/main.css
@@ -1,3 +1,0 @@
-*{
-	background-color: grey;
-}
