--- atlasDB/dashboard/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path, include
-from . import views
-
-urlpatterns = [
-	path('', views.dashboard, name='dashboard'),
-	path('aufgabe/<int:id>/', views.aufgabe, name='aufgabe'),
-]
